+++ src/Sniffs/CodeElement/FqcnMethodSniff.php
@@ -34,10 +34,7 @@ class FqcnMethodSniff implements CodeElementSniffInterface
     protected $reportMissingTags = true;
 
     /** @var bool */
-    protected $reportNullableBasicGetterDocType = true;
-
-    /** @var bool */
-    protected $reportNullableBasicGetterFnType = true;
+    protected $reportNullableBasicGetter = true;
 
     /**
      * @inheritDoc
@@ -53,8 +50,7 @@ class FqcnMethodSniff implements CodeElementSniffInterface
 
         $this->invalidTags = $invalidTags;
         $this->reportMissingTags = $config['reportMissingTags'] ?? true;
-        $this->reportNullableBasicGetterDocType = $config['reportNullableBasicGetterDocType'] ?? true;
-        $this->reportNullableBasicGetterFnType = $config['reportNullableBasicGetterFnType'] ?? true;
+        $this->reportNullableBasicGetter = $config['reportNullableBasicGetter'] ?? true;
     }
 
     /**
@@ -119,8 +115,7 @@ class FqcnMethodSniff implements CodeElementSniffInterface
             $this->processSigType($file, $subject);
 
             if ($method instanceof ClassMethodElement && $parent instanceof ClassElement) {
-                ($this->reportNullableBasicGetterDocType || $this->reportNullableBasicGetterFnType) // skip if both disabled
-                && static::reportNullableBasicGetter($file, $subject, $method, $parent);
+                $this->reportNullableBasicGetter && static::reportNullableBasicGetter($file, $subject, $method, $parent);
             }
         } else {
             foreach ($docBlock->getDescriptionLines() as $lineNum => $descLine) {
@@ -218,7 +213,7 @@ class FqcnMethodSniff implements CodeElementSniffInterface
         }
     }
 
-    protected function reportNullableBasicGetter(
+    protected static function reportNullableBasicGetter(
         File $file,
         ReturnTypeSubject $subject,
         ClassMethodElement $method,
@@ -249,7 +244,7 @@ class FqcnMethodSniff implements CodeElementSniffInterface
         $returnDocType = $subject->getDocType();
         $isGetterDocTypeNullable = TypeHelper::containsType($returnDocType, NullType::class);
         if (!$isGetterDocTypeNullable && $subject->hasDefinedDocBlock()) {
-            $this->reportNullableBasicGetterDocType && $subject->addDocTypeWarning(sprintf(
+            $subject->addDocTypeWarning(sprintf(
                 'Returned property $%s is nullable, add null return doc type, e.g. %s',
                 $propName,
                 ($returnDocType->toString() ?? $propDocType).'|null'
@@ -259,7 +254,7 @@ class FqcnMethodSniff implements CodeElementSniffInterface
         // Only report in fn type is defined. Doc type and fn type is synced by other sniffs.
         $returnFnType = $subject->getFnType();
         if (!($returnFnType instanceof UndefinedType) && !($returnFnType instanceof NullableType)) {
-            $this->reportNullableBasicGetterFnType && $subject->addFnTypeWarning(sprintf(
+            $subject->addFnTypeWarning(sprintf(
                 'Returned property $%s is nullable, use nullable return type declaration, e.g. ?%s',
                 $propName,
                 $returnFnType->toString()