+++ CHANGELOG.md
@@ -6,7 +6,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 
 ## 0.14.0 - 2020-??-??
 ### Added
-- Warning for uninitialized class/trait property. Default value and assignments in `__construct()` are checked. Add null doc type is suggested.
+- Warning for uninitialized class/trait property. Default value and assignments in `__construct()` are checked. Adding null doc type is suggested.
 - Warning for nullable return types for basic getters based on property doc type.
 - Setting `FqcnPropSniff.reportUninitializedProp`
 - Setting `FqcnPropSniff.reportNullableBasicGetter`